Updating helpfile about permanent locking
It looks like there is no graceful way to fit
a "manual unlock" link into the Moodle 1.9 gradebook,
so I'm just going to remove the wording that references
it. If teachers do need to manually unlock pages, they
can use the workaround of switching the assignment to
"unlock after grading" and touching all the grades.